lib/requirements.txt
docker
email_validator
datadog
bs4
flask
requests
google-api-python-client
google-cloud-datastore
google-cloud-storage
google-cloud-kms
pyyaml
iso8601
pytz
markdown2
flake8
